restructure: pages/app 폴더 정리
