--- src/app/pages/app/.gitkeep
@@ -1,1 +1,0 @@
-//빈 폴더 유지를 위한 파일
