+++ src/app/providers/index.tsx
@@ -2,7 +2,7 @@ import { PropsWithChildren } from 'react';
 
 import ReduxProvider from './ReduxProvider';
 import ReactQueryProvider from './ReactQueryProvider';
-import ToastContainer from '../shared/toast/ToastContainer';
+import ToastContainer from '@/shared/toast/ToastContainer';
 
 export default function Provider({ children }: PropsWithChildren) {
   return (
