+++ src/components/all-reviews/ReviewContainer.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useReviews } from '@/hooks/useReviews';
-import FilterBar from '@/components/shared/FilterBar';
+import FilterBar from '@/shared/FilterBar';
 import ReviewList from '@/components/all-reviews/ReviewList';
 import { FilterState } from '@/shared/types/reviewType';
 import Loader from './Loader';
