+++ src/pages/_app.tsx
@@ -3,7 +3,7 @@ import { SpeedInsights } from '@vercel/speed-insights/next';
 import { Analytics } from '@vercel/analytics/react';
 import Head from 'next/head';
 import GNB from '@/components/gnb';
-import Provider from '@/components/provider';
+import Provider from '@/app/providers';
 import '@/styles/globals.css';
 
 export default function App({ Component, pageProps }: AppProps) {
