+++ src/pages/app/lightning.tsx
@@ -1,4 +1,4 @@
-export { MyMeeting as default } from '@/app/pages/app/user/meeting';
+export { LightningPage as default } from '@/app/pages/app/lightning';
 
 // 기존 export default를 export로 변경했습니다.
