+++ src/pages/app/saved-gathering.tsx
@@ -1,4 +1,4 @@
-export { MyMeeting as default } from '@/app/pages/app/user/meeting';
+export { SavedGatheringPage as default } from '@/app/pages/app/saved-gathering';
 
 // 기존 export default를 export로 변경했습니다.
