+++ src/pages/app/start.tsx
@@ -1,4 +1,4 @@
-export { MyMeeting as default } from '@/app/pages/app/user/meeting';
+export { StartPage as default } from '@/app/pages/app/start';
 
 // 기존 export default를 export로 변경했습니다.
